Create Standart Tests

+++ scripts/update_doxygen.sh
@@ -2,23 +2,36 @@
 #
 # NO ARGUMENTS: 
 #
-#
-mkdir -p docs/
-#
 # STRING NOW
 now=$(date "+%Y/%m/%d - %H:%M:%S")
 #
 #
-echo "### START UPDATE DOXYGEN AT $now ###" >> outputs/update_oxygen.txt
+echo "### START UPDATE DOXYGEN AT $now ###"
 #
 #
 cd src/
 #
 #
-doxygen ../docs/doxygen_modem_on_config &>> ../outputs/update_oxygen.txt
+doxygen ../docs/doxygen_modem_on_config
+#
+# CATCH RESULT
+result=$?
+#
+# TEST RESULT
+if [ "$result" -ne "0" ]; then
+	#
+	#
+	echo "update_oxygen FAIL: $result"
+	#
+	#
+	cd ..
+	#
+	#
+	exit $result
+fi
 #
 #
-echo "### FINISH DOXYGEN ###" >> ../outputs/update_oxygen.txt
+echo "### FINISH UPDATE DOXYGEN ###"
 #
 # REMOVE OLD html/ AT docs/
 rm -rf ../docs/html
@@ -32,6 +45,8 @@ rm -rf latex/
 #
 cd ..
 #
+# 
+exit 0
 #
 # END OF SCRIPT
